gepeto_ai/requirements.txt
python-dotenv
langchain
openai
anthropic
google-generativeai
deepseek-ai
